+++ src/lib/typography/typography_caption_small.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 
 // Internal Components
-import TypographyBase from './typography_base';
+import TypographyBase, { TypographyColorType, TypographyFontFamily, TypographyMargin, TypographyTextAlign } from './typography_base';
 
 // Types
 import { ColorOptionType, FontWeightType } from '../private/types';
@@ -14,10 +14,12 @@ export default function TypographyCaptionSmall (props: {
     dataTestID?: string,
     ariaLabel?: string,
     color?: ColorOptionType,
-    margin?: "0" | "5" | "10" | "20" | "half" | "one" | "two",
+    margin?: TypographyMargin,
     fontWeight?: FontWeightType,
     style?: React.CSSProperties,
-    colorType?: 'normal' | 'text' | 'light',
+    colorType?: TypographyColorType,
+    fontFamily?: TypographyFontFamily,
+    textAlign?: TypographyTextAlign
 }) {
     return <TypographyBase 
         text={props.text}
@@ -30,5 +32,7 @@ export default function TypographyCaptionSmall (props: {
         fontWeight={props.fontWeight}
         style={props.style}
         colorType={props.colorType}
+        fontFamily={props.fontFamily}
+        textAlign={props.textAlign}
     />
 }