+++ src/view/typography/typography.tsx
@@ -32,11 +32,24 @@ export function typographyOptions(): ViewData {
             },
             showTitle: 'booleanTrueDefault',
             color: 'colorsOptional',
+            colorType: {
+                options: ["normal", 'text', "light", "text-light"],
+                default: "normal"
+            },
             margin: {
                 default: '0',
                 options: ["0", "5", "10", "20", "half", "one", "two"]
             },
-            fontWeight: 'fontWeightOptional'
+            fontWeight: 'fontWeightOptional',
+            fontFamily: {
+                options: ["primary", "secondary"],
+                default: "primary"
+            },
+            textAlign: {
+                options: ["left", "center", "right", "justify"],
+                default: "left"
+            },
+            italic: 'boolean'
         }
     }
 }
@@ -49,7 +62,11 @@ export function TypographyCreator(props: { p: TypographyPropsType }) {
         text='Sample Text'
         showTitle={props.p.showTitle}
         color={props.p.color}
+        colorType={props.p.colorType}
         margin={props.p.margin}
         fontWeight={props.p.fontWeight}
+        fontFamily={props.p.fontFamily}
+        style={{fontStyle: (props.p as any).italic ? "italic" : "normal"}}
+        textAlign={props.p.textAlign}
     />
 }